delivering for my precious testers

+++ backend/api/controllers/labsController.go
@@ -32,6 +32,16 @@ func LabsGet(c *fiber.Ctx) error {
 }
 func LabDelete(c *fiber.Ctx) error {
 	id := c.Params("id")
+
+	userCredentials := c.Locals("user").(fiber.Map)
+	role := userCredentials["Role"].(string)
+
+	if role != "tutor" {
+		return c.Status(http.StatusUnauthorized).JSON(fiber.Map{
+			"error": "unauthorized to perform this action",
+		})
+	}
+
 	result := initializers.DB.Delete(&models.Lab{}, id)
 	if result.Error != nil {
 		return c.Status(http.StatusBadRequest).JSON(
